Move wsgi_mkenv from haas.rest to haas.test_common
We're going to use this to fabricate http requests in tests yet to be
written, so it needs to be somewhere where it can be imported.
